requirements.txt
langchain
python-dotenv
langchain[groq]
langchain-core
langgraph>0.2.27
ddgs
langchain-community